creating symlink for files instead of coping

+++ setup_dotfile.py
@@ -1,5 +1,5 @@
 import os
-import shutil
+#import shutil
 from datetime import datetime
 from pathlib import Path
 
@@ -21,10 +21,11 @@ def setup_dotfiles():
     for file_name in file_names:
         full_file_path = file_names[file_name] + file_name
 
-        if os.path.exists(full_file_path):
+        if os.path.exists(full_file_path) and not os.path.islink(full_file_path):
             backup_file_name = f"{full_file_path}-{run_time}"
             print(f"Saving a backup in: {backup_file_name}")
             os.rename(full_file_path, backup_file_name)
 
-        print(f"Copying file ./dotfiles/{file_name} to {full_file_path}")
-        shutil.copy(f"./dotfiles/{file_name}", f"{full_file_path}")
+        print(f"Creating symlink for ./dotfiles/{file_name} in {full_file_path}")
+        os.symlink(f'./dotfiles/{file_name}', f'{full_file_path}')
+        #shutil.copy(f"./dotfiles/{file_name}", f"{full_file_path}")
